Create separate CUDA Engine for generating and executing PTX
