+++ include/reffine/engine/engine.h
@@ -50,9 +50,6 @@ public:
 
     static ExecEngine* Get();
     void Optimize(Module&);
-    void GeneratePTX(Module&, std::string&);
-    void ExecutePTX(const std::string&, const std::string&, void*, int*);
-    void ExecutePTXFromFile(const std::string&, const std::string&, void*, int*);
     void AddModule(unique_ptr<Module>);
     LLVMContext& GetCtx();
